Fix: FolderPage issue
Fixed issue where the FolderPage quartz plugin would fail because the stack size was exceeded.

It seems that the problem was caused by the following line from `content/Library/The Bible (WEB)/00 - About/God in the Bible/attributes of God/index.md`:
```
> ![[Library/The Bible (WEB)/00 - About/God in the Bible/attributes of God/index|📂 Folder Note Map]]
```

I think that this line was causing an infinite regression.